Renamed study util js files to match variable names
